[Silver III] Title: 피보나치는 지겨웡~, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/17175. 피보나치는 지겨웡～/피보나치는 지겨웡～.cc
@@ -8,7 +8,7 @@
 using namespace std;
 
 int n;
-long long int dp[51];
+int dp[51];
 
 int main()
 {
